Modifying water alerts to include alert for supervisors. Moving functions to base notification
