+++ WNPRC_EHR/src/org/labkey/wnprc_ehr/notification/WaterMonitoringAnimalWithOutEntriesNotification.java
@@ -40,7 +40,7 @@ import java.util.List;
 import java.util.Map;
 
 
-public class WaterMonitoringAnimalWithOutEntriesNotification extends AbstractEHRNotification
+public class WaterMonitoringAnimalWithOutEntriesNotification extends WaterMonitoringNotification
 {
     public WaterMonitoringAnimalWithOutEntriesNotification(Module owner)
     {
@@ -49,7 +49,7 @@ public class WaterMonitoringAnimalWithOutEntriesNotification extends AbstractEHR
 
     public String getName()
     {
-        return "Water Monitoring";
+        return "Water Monitoring Animal For Vets and Lab";
     }
 
 
@@ -62,10 +62,6 @@ public class WaterMonitoringAnimalWithOutEntriesNotification extends AbstractEHR
 
     @Override
     public String getCronString() { return "0 0 15,19 * * ?"; }
-
-    public String getCategory(){
-        return "Husbandry";
-    }
 
     @Override
     public String getScheduleDescription()
@@ -82,127 +78,24 @@ public class WaterMonitoringAnimalWithOutEntriesNotification extends AbstractEHR
     public String getMessageBodyHTML(final Container c, User u)
     {
         final StringBuilder msg = new StringBuilder();
+        int numDays = 5;
 
         //Find today's date
         Date now = new Date();
         msg.append("This email contains a series of automatic alerts about the water monitoring system.  It was run on: " + AbstractEHRNotification._dateFormat.format(now) + " at " + AbstractEHRNotification._timeFormat.format(now) + ".<p>");
 
+        //Check animals that did not get any water for today and the last five days.
+        findAnimalsWithWaterEntries(c,u,msg,numDays);
+
         //Check animals with less than 20 mls per kilogram of water for today
         findAnimalsWithEnoughWater(c,u,msg);
 
-        //Check animals that did not get any water for today and the last five days.
-        findAnimalsWithWaterEntries(c,u,msg);
 
 
         return msg.toString();
     }
 
-    protected void findAnimalsWithEnoughWater(final Container c, final User u, final StringBuilder msg)
-    {
-        Calendar cal = Calendar.getInstance();
-        cal.setTime(new Date());
 
-
-        SimpleFilter.OrClause orClause = new SimpleFilter.OrClause();
-        orClause.addClause(new SimpleFilter(FieldKey.fromString("mlsPerKg"), 20, CompareType.LT).getClauses().get(0));
-        orClause.addClause(new SimpleFilter(FieldKey.fromString("mlsPerKg"),null, CompareType.ISBLANK).getClauses().get(0));
-        SimpleFilter filter = new SimpleFilter(FieldKey.fromString("date"), cal.getTime(), CompareType.DATE_EQUAL);
-        filter.addClause(orClause);
-
-        TableSelector ts = new TableSelector(getStudySchema(c, u).getTable("waterTotalByDateWithWeight"),PageFlowUtil.set("Id","date","mlsPerKg","TotalWater","project"), filter, null);
-        long count = ts.getRowCount();
-        if (count > 0)
-        {
-            msg.append("<b>WARNING: There are " + count + " animals that have remaining water for today.</b><br>\n");
-            Map<String,Object>[] totalWaterForDay = ts.getMapArray();
-            msg.append("<table border=1 style='border-collapse: collapse;'>");
-            msg.append("<tr><td style='padding: 5px; text-align: center;'><strong>Project</strong></td>" +
-                    "<td style='padding: 5px; text-align: center;'><strong>Id</strong></td>" +
-                    "<td style='padding: 5px; text-align: center;'><strong>Date</strong></td>" +
-                    "<td style='padding: 5px; text-align: center;'><strong>mlsPerKg</strong></td>" +
-                    "<td style='padding: 5px; text-align: center;'><strong>Total Water Given</strong></td></tr>\n");
-
-            Map<Integer, List<Map<String,Object>>> projectMap = new HashMap<>();
-            for(Map<String,Object> mapItem : totalWaterForDay){
-                int projectNum = ConvertHelper.convert(mapItem.get("project"),Integer.class);
-                List<Map<String,Object>> waterTotalsFromDb;
-                if (!projectMap.containsKey(projectNum)){
-                    waterTotalsFromDb = new ArrayList<>();
-                    projectMap.put(projectNum,waterTotalsFromDb);
-                }else{
-                    waterTotalsFromDb = projectMap.get(projectNum);
-                }
-                waterTotalsFromDb.add(mapItem);
-            }
-
-            for (Map.Entry<Integer,List<Map<String,Object>>> entry : projectMap.entrySet()){
-                List<Map<String,Object>> totalWaterByProject = entry.getValue();
-                String mlsPerKg;
-                String totalWater;
-                for(Map<String,Object> mapItem : totalWaterByProject){
-                    LocalDateTime objectDateTime = ConvertHelper.convert(mapItem.get("date"),Date.class).toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime();
-                    DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-
-                     mlsPerKg = ConvertHelper.convert(mapItem.get("mlsPerKg"),String.class) == null ? " " : ConvertHelper.convert(mapItem.get("mlsPerKg"),String.class);
-                     totalWater = ConvertHelper.convert(mapItem.get("TotalWater"),String.class) == null ? " " : ConvertHelper.convert(mapItem.get("TotalWater"),String.class);
-
-                    msg.append("<tr><td style='padding: 5px;'>" + ConvertHelper.convert(mapItem.get("project"),Integer.class)
-                            + "</td><td style='padding: 5px; text-align: center;'> " + ConvertHelper.convert(mapItem.get("Id"),String.class)
-                            + "</td><td style='padding: 5px; text-align: center;'> " + objectDateTime.format(formatter)
-                            + "</td><td style='padding: 5px; text-align: center;'> " + mlsPerKg
-                            + "</td><td style='padding: 5px; text-align: center;'> " + totalWater
-                            +"</td></tr>" );
-
-                }
-            }
-
-
-
-
-            msg.append("</table>");
-            msg.append("<p><a href='" + getExecuteQueryUrl(c, "study", "waterTotalByDateWithWeight", null) + "&query.date~dateeq=" + AbstractEHRNotification._dateFormat.format(cal.getTime()) +"&query.mlsPerKg~lt=20'>Click here to view them</a><br>\n\n");
-            msg.append("<hr>\n\n");
-
-        }
-    }
-
-    protected  void findAnimalsWithWaterEntries(final Container c, final User u, final StringBuilder msg){
-
-        LocalDate date = LocalDate.now();
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-        int datesInPast = 5;
-
-        for (int i = 0; i<datesInPast; i++){
-
-            Map<String, Object> parameters = new CaseInsensitiveHashMap<>();
-            parameters.put("CheckDate", date);
-
-            TableInfo ti = QueryService.get().getUserSchema(u, c, "study").getTable("waterScheduledAnimalWithOutEntries");
-            TableSelector ts = new TableSelector(ti, PageFlowUtil.set("id","project"), null, null);
-            ts.setNamedParameters(parameters);
-
-            long total = ts.getRowCount();
-
-            if (total == 0)
-            {
-                msg.append("All regulated animals have at least one entries for "+ date.getDayOfWeek().toString() +" ("+ date.format(formatter) + ").<br>");
-            }
-            else
-            {
-                msg.append("There are " + total + " animals in the system that have no records in water given dataset for " + date.getDayOfWeek().toString() +" ("+ date.format(formatter) + ").<br>");
-                Map<String, Object>[] animalsWithOutEntries = ts.getMapArray();
-                for (Map<String, Object> mapItem : animalsWithOutEntries)
-                {
-                    msg.append(ConvertHelper.convert(mapItem.get("project"), Integer.class) + "  "  + ConvertHelper.convert(mapItem.get("Id"), String.class) + "<br>");
-
-                }
-                msg.append("<br>");
-            }
-            date = date.minusDays(1);
-        }
-
-
-    }
 
 
 }
